ENH: reowrk search components folder structure
