+++ src/app/components/Search/PlayerResults/styled.js
