+++ src/app/components/Search/Result404/index.js
