+++ src/app/components/Search/Result404/styled.js
