+++ src/app/components/Search/SearchTabs/SearchContent/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import MatchResults from 'app/components/MatchResults'
-import PlayerResults from 'app/components/PlayerResults'
+import MatchResults from 'app/components/Search/MatchResults'
+import PlayerResults from 'app/components/Search/PlayerResults'
 
 const renderContent = selected => {
   switch (selected) {
