+++ src/app/components/TeamHeroes/styled.js
